chore: simplified QueryInvalidInput to match new quaint edits
