+++ query-engine/connector-test-kit-rs/query-engine-tests/tests/new/mod.rs
@@ -1,8 +1,8 @@
+mod assertion_violation_error;
 mod create_many;
 mod cursor;
 mod disconnect;
 mod interactive_tx;
-mod invalid_input_error;
 mod metrics;
 mod multi_schema;
 mod native_types;
